made metadata file create/git-add more consistent
May not be the best solution.  When files are not created this is
noted as an error in logs, and files are not added to the
list of files to be added.  The commit can then go through bc
all the files that were present were staged and committed, when
in fact some files were missing.  This should throw an error.
But my eyes are tired, I'll have to come back to it.