commands return status,message; CLI wrapper returns sys.exit
I want the command functions to be usable within a Python app.
Therefore they cannot be doing sys.exit(1) which would halt
execution of the app that employed them.
Modded the functions to return a tuple consisting of an exit
status (0 or 1) and a string message ('ok' or an error message).
CLI wrapper returns sys.exit status based on what comes out
of the commands.