replace unittest framework with roll-yer-own TAP
unittest is great for testing code but isn't flexible enough for
this project.  In particular I needed to be able to return lists
of files/entities/things that don't pass tests, and the output
needs to be parsable by other programs.
TAP (Test Anything Protocol) seems like a good candidate.
The new code should make it easier to return a list of tests and
results when called as a library, and nicely formatted text when
run from the command line.