add/update entity json,mets on entity-annex-add but don't edit
The contents of entity.json and mets.xml were being changed by
both ddr and ddrlocal, which was causing some data to be lost
(file role field).  This change includes entity.json and mets.xml
in the commit when adding a file to an entity but does not
mess with the contents of these files, which should henceforth
be managed by ddrlocal.