rm JSON,XML processing code; rm DDRCollection,DDREntity
JSON and XML metadata files are best handled at higher layers in
the DDR software stack.  This change removes the rest of the code
that deals with JSON and XML, as well as template files.

The DDRCollection and DDREntity classes' only function was to
add metadata to Collection and Entity, so these classes have also
been removed.