various experiments in backup/cloning of groups of repos
My idea at this point is to have an "organization" directory that
contains metadata about the organization, plus a number of
"group files" with metadata for groups of repos.
Groups may represent physical drives or simply logical groups.
Groups are simply groups of repos that are located together and
which share the same "level", or percentage of annex-managed
binary files that are present in the repo's annex.
The idea is to make it easier to backup/clone many repos at once.

- Moved git functions from DDR.commands to DDR.dvcs.
- Added organization module with functions for reading/writing
  mrconfig files and groupfiles.
- Added some commands to DDR.commands and bin/ddr to try from CLI.
- Shell script that uses mr and ddr commands.
- Python script that does away with mr.
issue #68