replace git-annex-sync substitute with the real git-annex-sync
Repository create and sync functions used the following sequence
of commands instead of "git annex sync", This because of fears
that "annex sync" might be doing unwanted operations.
  git fetch origin
  git checkout master
  git pull origin master
  git checkout git-annex
  git pull origin git-annex
  git push origin git-annex
  git checkout master
  git push origin master

These fears appear to be unfounded.  Also, the above sequence
appears to no longer be working, making it impossible to create
new collections or sync existing ones.  It was replaced with
"git annex sync" and now create and sync appear to work correctly.